cfx_storage::state

Struct State

pub struct State { /* private fields */ }

A block defines a list of transactions that it sees and the sequence of the transactions (ledger). At the view of a block, after all transactions being executed, the data associated with all addresses is a State after the epoch defined by the block.

A writable state is copy-on-write reference to the base state in the state manager. State is supposed to be owned by single user.

Implementations

impl State

pub fn new( manager: Arc<StateManager>, state_trees: StateTrees, construct_pivot_state: bool, ) -> Self

pub fn get_from_snapshot<WithProof: StaticBool>( &self, access_key: &[u8], ) -> Result<(Option<Box<[u8]>>, Option<TrieProof>)>

impl State

pub fn dump<DUMPER: KVInserter<MptKeyValue>>( &self, dumper: &mut DUMPER, ) -> Result<()>