Trait cfx_storage::storage_db::key_value_db::KeyValueDbAsAnyTrait

pub trait KeyValueDbAsAnyTrait: KeyValueDbTypes {
    // Required method
    fn as_any(&self) -> &dyn Any;
}

This trait is to help with the committing of the transaction for which the db object should be provided for serialization.

Required Methods

fn as_any(&self) -> &dyn Any