cfx_storage

Struct SqliteConnection

pub struct SqliteConnection { /* private fields */ }

Implementations

impl SqliteConnection

pub fn close(&mut self) -> Result<()>

pub fn close_v2(&mut self) -> Result<()>

pub fn default_open_flags() -> OpenFlags

pub fn create_and_init<P: AsRef<Path>>(path: P, unsafe_mode: bool) -> Result<()>

If unsafe_mode is true, data loss or database corruption may happen if the process crashes, so it should only be used for write-once databases where an unfinished temporary database will be removed after process restart.

pub fn create_and_open<P: AsRef<Path>>( path: P, open_flags: OpenFlags, unsafe_mode: bool, ) -> Result<Self>

pub fn open<P: AsRef<Path>>( path: P, readonly: bool, open_flags: OpenFlags, ) -> Result<Self>

pub fn try_clone(&self) -> Result<Self>

pub fn prepare<'db>( db: &'db mut Connection, statement_cache: &'db mut HashMap<String, ScopedStatement>, sql: &str, ) -> Result<&'db mut ScopedStatement>

pub fn execute_locked<'db, 'p, Param: Borrow<dyn SqlBindable + 'p>>( statement: &'db mut ScopedStatement, params: &[Param], ) -> Result<MaybeRows<'db>>

The statement must be created with the db. Then the statement is a mut borrow of db, so it’s guaranteed that the db is only used by one thread.

pub fn execute<'p, Param: Borrow<dyn SqlBindable + 'p>>( &mut self, sql: &str, params: &[Param], ) -> Result<MaybeRows<'_>>

pub fn get_db_mut(&mut self) -> &mut Connection

pub fn lock_db(&self) -> MutexGuard<'_, Connection>

pub fn lock_statement_cache( &self, ) -> MutexGuard<'_, HashMap<String, ScopedStatement>>

pub fn possible_temporary_files(db_path: &str) -> Vec<String>