cfx_storage::tests

Struct FakeSnapshotMptDb

pub struct FakeSnapshotMptDb { /* private fields */ }

Implementations

impl FakeSnapshotMptDb

pub fn new_discard_write() -> Self

With discard_write, it tests if the MptMerger does not rely on the value to write to work correctly. In other words, MptMerger works correctly only reading the data to merge.