Module cfx_storage::utils

Modules

access_mode

deref_plus_impl_or_borrow_self

guarded_value

tuple

wrap

Traits

StateRootWithAuxInfoToFromRlpBytes

Only used by storage benchmark due to incompatibility of rlp crate version.

UnsafeCellExtension

WrappedCreateFrom

The purpose of this trait is to create a new value of a passed object, when the passed object is the value, simply move the value; when the passed object is the reference, create the new value by clone. Other extension is possible.

Functions

to_key_prefix_iter_upper_bound