cfx_storage::utils::deref_plus_impl_or_borrow_self

Trait ImplOrBorrowSelf

pub trait ImplOrBorrowSelf<Trait: ?Sized> {
    // Required method
    fn borrow(&self) -> &Trait;
}

This trait is automatically implemented for the Trait ifself, any type which implements Trait, and any Borrow of the Trait.

Required Methods

fn borrow(&self) -> &Trait