Struct cfx_storage::utils::guarded_value::GuardedValue

pub struct GuardedValue<GuardType, ValueType> { /* private fields */ }

Please read all comments in this file to prevent leaking of guarded value by accident.

Implementations

impl<GuardType, ValueType> GuardedValue<GuardType, ValueType>

pub fn new(guard: GuardType, value: ValueType) -> Self

pub fn into(self) -> (GuardType, ValueType)

It’s unsafe to drop the returned guard manually, or to clone the ValueType. Extra care is needed when using this method.

impl<'a, GuardType: 'a + Deref<Target = TargetType>, TargetType> GuardedValue<GuardType, NonCopy<&'a TargetType>>

pub fn new_derefed(guard: GuardType) -> Self

impl<'a, GuardType: 'a + DerefMut<Target = TargetType>, TargetType> GuardedValue<GuardType, &'a mut TargetType>

pub fn new_derefed_mut(guard: GuardType) -> Self

impl<'a, GuardType> GuardedValue<GuardType, MaybeOwnedTrieNode<'a>>

pub fn take( x: Self ) -> GuardedValue<GuardType, MaybeOwnedTrieNodeAsCowCallParam>

impl<'a, GuardType> GuardedValue<GuardType, &'a TrieNodeDeltaMptCell>

pub fn into_wrapped(x: Self) -> GuardedValue<GuardType, MaybeOwnedTrieNode<'a>>