cfx_storage::utils

Module tuple

Re-exports

pub use placeholders::PLACEHOLDERS;

Modules

placeholders

Structs

ElementNoConstrain

Traits

ElementConstrainMark

To make it easy for compiler to compute automated implementation.

ElementSatisfy

ElementConstrain is a trait that can be made into trait object where the tuple element satisfies.

IndexGetExt

IterCallFamilyTrait

OfElementSatisfiesOnTuple

Trait on Index, meaning that the element at Index for ConcernedTuple satisfies ElementConstrain.

TestTupleIndexExt

TupleGetIndexExt

This trait is on Tuple.

TupleIndex

TupleIndexExt

TupleIndexUpTo

TupleIterFromTo

TupleIterate

TupleIterateFromTo