cfx_storage::storage_db

Module key_value_db

Structs

KvdbIterIterator

Traits

DbImplByFamily

DbImplFamily

DbValueType

One of the elementary value type supported by db.

KeyValueDbAsAnyTrait

This trait is to help with the committing of the transaction for which the db object should be provided for serialization.

KeyValueDbIterableTrait

KeyValueDbToOwnedReadTrait

KeyValueDbTrait

KeyValueDbTraitMultiReader

The difference between this trait and KeyValueDbTraitRead is that, the type which implements KeyValueDbTraitRead may use lock to serialize the reads, which is not necessarily a multi reader.

KeyValueDbTraitOwnedRead

These special get methods are provided for db like sqlite, where concurrency can only be achieved by opening a separate connection, otherwise lock is required for concurrent read.

KeyValueDbTraitRead

KeyValueDbTraitSingleWriter

KeyValueDbTraitSingleWriterMultiReader

KeyValueDbTraitTransactional

KeyValueDbTraitTransactionalDyn

KeyValueDbTransactionTrait

KeyValueDbTypes

OwnedReadImplByFamily

OwnedReadImplFamily

ReadImplByFamily

ReadImplFamily

SingleWriterImplByFamily

SingleWriterImplFamily