cfx_storage::storage_db::key_value_db

Trait KeyValueDbTraitOwnedRead

pub trait KeyValueDbTraitOwnedRead: KeyValueDbTypes {
    // Required method
    fn get_mut(&mut self, key: &[u8]) -> Result<Option<Self::ValueType>>;

    // Provided method
    fn get_mut_with_number_key(
        &mut self,
        key: i64,
    ) -> Result<Option<Self::ValueType>> { ... }
}

These special get methods are provided for db like sqlite, where concurrency can only be achieved by opening a separate connection, otherwise lock is required for concurrent read.

Required Methods

fn get_mut(&mut self, key: &[u8]) -> Result<Option<Self::ValueType>>

Provided Methods

fn get_mut_with_number_key( &mut self, key: i64, ) -> Result<Option<Self::ValueType>>